refactor: Schedule to calendar view, support decimal band targets, remove Goals page
- Refactor Schedule page to use react-big-calendar Month view
- Add custom CSS styles for calendar appearance
- Fix Settings to accept decimal band values (5.5, 6.5, 7.5, 8.5)
- Fix Skills page target display for decimal values
- Remove Goals page (redundant with Skills + Settings)
- Add react-big-calendar and date-fns dependencies

+++ src/pages/Settings.tsx
@@ -24,7 +24,7 @@ export default function Settings() {
   }, [skills, scores]);
 
   const handleTargetChange = (skillId: string, val: string) => {
-    const band = parseInt(val);
+    const band = parseFloat(val);
     if (band >= 1 && band <= 9) updateSkillTarget(skillId, band);
   };
